project strucute reorganization
