+++ src/main/java/com/rj/ecommerce_backend/GlobalExceptionHandler.java
@@ -1,19 +1,18 @@
 package com.rj.ecommerce_backend;
 
-import com.rj.ecommerce_backend.domain.order.exceptions.AccessDeniedException;
-import com.rj.ecommerce_backend.domain.order.exceptions.OrderCancellationException;
-import com.rj.ecommerce_backend.domain.order.exceptions.OrderNotFoundException;
-import com.rj.ecommerce_backend.domain.order.exceptions.OrderServiceException;
-import com.rj.ecommerce_backend.domain.product.dtos.ErrorDTO;
-import com.rj.ecommerce_backend.domain.product.exceptions.CategoryNotFoundException;
-import com.rj.ecommerce_backend.domain.product.exceptions.InsufficientStockException;
-import com.rj.ecommerce_backend.domain.product.exceptions.ProductNotFoundException;
-import com.rj.ecommerce_backend.domain.user.exceptions.UserNotFoundException;
+import com.rj.ecommerce_backend.order.exceptions.AccessDeniedException;
+import com.rj.ecommerce_backend.order.exceptions.OrderCancellationException;
+import com.rj.ecommerce_backend.order.exceptions.OrderNotFoundException;
+import com.rj.ecommerce_backend.order.exceptions.OrderServiceException;
+import com.rj.ecommerce_backend.product.dtos.ErrorDTO;
+import com.rj.ecommerce_backend.product.exceptions.CategoryNotFoundException;
+import com.rj.ecommerce_backend.product.exceptions.InsufficientStockException;
+import com.rj.ecommerce_backend.product.exceptions.ProductNotFoundException;
+import com.rj.ecommerce_backend.user.exceptions.UserNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.FieldError;
-import org.springframework.web.ErrorResponse;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
