+++ src/main/java/com/rj/ecommerce_backend/securityconfig/SecurityContext.java
@@ -1,6 +1,6 @@
 package com.rj.ecommerce_backend.securityconfig;
 
-import com.rj.ecommerce_backend.domain.user.User;
+import com.rj.ecommerce_backend.user.domain.User;
 
 public interface SecurityContext {
     void checkAccess(Long userId);
