+++ src/main/java/com/rj/ecommerce_backend/securityconfig/services/AuthService.java
@@ -1,6 +1,6 @@
 package com.rj.ecommerce_backend.securityconfig.services;
 
-import com.rj.ecommerce_backend.domain.user.User;
+import com.rj.ecommerce_backend.user.domain.User;
 import com.rj.ecommerce_backend.securityconfig.dto.AuthResponse;
 import com.rj.ecommerce_backend.securityconfig.dto.LoginRequest;
 import com.rj.ecommerce_backend.securityconfig.dto.TokenRefreshRequest;
